use ssize_t instead of loff_t as the former is portable and the latter is not
git-svn-id: svn://localhost/svn/toolchain/trunk@1097 2adf4935-75e9-4407-93b2-34b21a1880a4

+++ cramfs-1.1/mkcramfs.c
@@ -371,7 +371,7 @@ static int cramsort (const void *a, const void *b)
 		       (*(const struct dirent **) b)->d_name);
 }
 
-static unsigned int parse_directory(struct entry *root_entry, const char *name, struct entry **prev, loff_t *fslen_ub)
+static unsigned int parse_directory(struct entry *root_entry, const char *name, struct entry **prev, ssize_t *fslen_ub)
 {
 	struct dirent **dirlist;
 	int totalsize = 0, dircount, dirindex;
@@ -896,7 +896,7 @@ static struct entry *find_filesystem_entry(struct entry *dir, char *name, mode_t
 }
 
 void modify_entry(char *full_path, unsigned long uid, unsigned long gid, 
-	unsigned long mode, unsigned long rdev, struct entry *root, loff_t *fslen_ub)
+	unsigned long mode, unsigned long rdev, struct entry *root, ssize_t *fslen_ub)
 {
 	char *name, *path, *full;
 	struct entry *curr, *parent, *entry, *prev;
@@ -1030,7 +1030,7 @@ inline int snprintf(char *str, size_t n, const char *fmt, ...)
     block, fifo, or directory does not exist, it will be created.
 */
 
-static int interpret_table_entry(char *line, struct entry *root, loff_t *fslen_ub)
+static int interpret_table_entry(char *line, struct entry *root, ssize_t *fslen_ub)
 {
 	char type, *name = NULL;
 	unsigned long mode = 0755, uid = 0, gid = 0, major = 0, minor = 0;
@@ -1086,7 +1086,7 @@ static int interpret_table_entry(char *line, struct entry *root, loff_t *fslen_u
 	return 0;
 }
 
-static int parse_device_table(FILE *file, struct entry *root, loff_t *fslen_ub)
+static int parse_device_table(FILE *file, struct entry *root, ssize_t *fslen_ub)
 {
 	char *line;
 	int status = 0;
@@ -1186,7 +1186,7 @@ int main(int argc, char **argv)
 	ssize_t offset, written;
 	int fd;
 	/* initial guess (upper-bound) of required filesystem size */
-	loff_t fslen_ub = sizeof(struct cramfs_super);
+	ssize_t fslen_ub = sizeof(struct cramfs_super);
 	char const *dirname, *outfile;
 	u32 crc;
 	int c;			/* for getopt */
